+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This project won't have versionining per se. But it's format is based in the [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2022-07-25
+
+### Added
+
+- Created the github repo utility to check how many repositories did I created since 2021 (when I left my first job).
+
 ## 2022-07-12
 
 ### Added